Update download page to use stack unpack
This has two main benefits over what we have currently:

It means that people end up using the same install plan that we use in
the binaries; currently, encouraging people to use nightly resolvers
will mean that they sometimes get new versions of libraries we depend
on, resulting in a set of dependencies which we haven't run the compiler
test suite with, or even used together.

It also means that people aren't limited to only those versions of
purescript which have made it into a nightly version. For example, when
we depend on some other package which isn't yet in the nightly snapshots
(for example, because not everything in the snapshot works with its
newer versions yet), that version of purescript would not be
installable. By contrast, this approach will work for any version of
PureScript which is on Hackage.